handle marks on elements with content
This is an attempt that works in a few examples I tried out. For some reason trying to use rollup locally results in an unusable dist file, so I was forced to make the changes directly in the ES5 version dist folder and here I recreate what I did here. Someone with more knowledge of the sources should take a look at it. Relates to https://github.com/ProseMirror/prosemirror/issues/780 .